+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED (0.14.6)
+
+🗑️ **Deprecations**
+
+- Not passing a `$scale` to `BigDecimal::dividedBy()` is deprecated; **`$scale` will be required in 0.15**
+
 ## [0.14.5](https://github.com/brick/math/releases/tag/0.14.5) - 2026-02-03
 
 🗑️ **Deprecations**